backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
pydantic-settings>=2.5,<2.6
sqlalchemy>=2.0,<2.1
alembic>=1.13,<1.14
psycopg[binary]>=3.2,<3.3
pgvector>=0.3,<0.4
